requirements.txt
Flask>=3.0
python-binance
pandas
python-dotenv
